requirements.txt
web3
python-dotenv
aiohttp
websockets
redis
